+++ src/StateSmith/Input/antlr4/LabelParser.cs
@@ -64,12 +64,15 @@ namespace StateSmith.Input.antlr4
         {
             ICharStream stream = CharStreams.fromString(inputString);
             var lexer = new StateSmithLabelGrammarLexer(stream);
+            lexer.RemoveErrorListeners(); // prevent antlr4 error output to console
             lexer.AddErrorListener(errorListener);
+
             ITokenStream tokens = new CommonTokenStream(lexer);
             StateSmithLabelGrammarParser parser = new StateSmithLabelGrammarParser(tokens)
             {
                 BuildParseTree = true
             };
+            parser.RemoveErrorListeners(); // prevent antlr4 error output to console
             parser.AddErrorListener(errorListener);
 
             return parser;